+++ react/health-map/CHANGELOG.md
@@ -1,7 +1,10 @@
 # CHANGELOG
 
 All notable changes to this project will be documented in this file.
-CURRENT VERSION: 2.0.0 - FORMAT: version - dd/mm/yyyy
+CURRENT VERSION: 3.0.0 - FORMAT: version - dd/mm/yyyy
+
+## 3.0.0 - 20/04/2020
+Support of geo points with cards on click
 
 ## 2.0.0 - 03/04/2019
 Guayaquil Health Map 2.0.0